requirements.txt
pocketflow>=0.0.1
google-generativeai>=0.3.0
gitpython>=3.1.0
openai>=1.0.0